+++ rollup.config.js
@@ -7,7 +7,7 @@ import typescript from '@rollup/plugin-typescript';
 import peerDeps from 'rollup-plugin-peer-deps-external';
 import postcss from 'rollup-plugin-postcss';
 import visualizer from 'rollup-plugin-visualizer';
-import { terser } from 'rollup-plugin-terser';
+// import { terser } from 'rollup-plugin-terser';
 
 export default {
 	input: './src/index.ts',
@@ -31,11 +31,11 @@ export default {
 			minimize: true,
 			extract: true
 		}),
-		terser(),
+		// terser(),
 		visualizer({
 			filename: 'bundle-analysis.html',
 			open: true
 		})
 	],
-	external: ['classnames']
+	external: ['classnames', 'tslib', 'lodash.clonedeep', 'lodash.clone']
 };