+++ src/ui/index.tsx
@@ -1,7 +1,29 @@
-export { default as Label } from './label/Label';
-export { default as Button } from './button/Button';
-export { default as Box } from './box/Box';
+// We have to separate the Props and Component into different exports see this issue:
+// https://github.com/vitejs/vite/issues/2117
+export { Label } from './label/Label';
+export type { LabelProps } from './label/Label';
+
+export { Button } from './button/Button';
+export type { ButtonProps } from './button/Button';
+
+export { Box } from './box/Box';
+export type { BoxProps } from './box/Box';
+
 export { Icon } from './icon/Icon';
+export type { IconProps } from './icon/Icon';
+
 export { InputText } from './inputText/InputText';
 export { InputTextarea } from './inputTextarea/InputTextarea';
+export type { InputTextProps } from './inputText/InputText';
+
 export { Chip } from './chip/Chip';
+export type { ChipProps } from './chip/Chip';
+
+export { popupController } from './popupController/popupController';
+
+export { Popup } from './popupController/Popup';
+export type { PopupProps } from './popupController/Popup';
+
+export { RsFormControl } from './form/FormControl';
+export { RsFormGroup } from './form/FormGroup';
+export { RsValidator, RsValidatorEnum } from './form/Validator';