+++ src/ui/inputText/InputText.tsx
@@ -3,22 +3,16 @@ import './InputText.scss';
 import { InputHTMLAttributes, useEffect, useRef } from 'react';
 import classNames from 'classnames';
 import { ICommon } from '../../common/Interfaces';
-import { RsFormControl } from '../form';
-import { IRsFormControl } from '../form/FormControl';
+import { RsFormControl, IRsFormControl } from '../form/FormControl';
 
-import Box from '../box/Box';
+import { Box } from '../box/Box';
 import clone from 'lodash.clone';
-import { Icon, IconProps } from '../icon/Icon';
+import { Icon } from '../icon/Icon';
 
-export interface NewIconProps extends IconProps {
-	position: 'LEFT' | 'RIGHT';
-	isHidden?: boolean;
-}
-
-interface InputTextProps extends Omit<InputHTMLAttributes<HTMLInputElement>, 'onChange' | 'type'> {
+export interface InputTextProps
+	extends Omit<InputHTMLAttributes<HTMLInputElement>, 'onChange' | 'type'>,
+		Omit<ICommon.HtmlElementProps, 'display'> {
 	//TextInput Props
-	id?: string;
-	className?: string;
 	boxRef?: React.RefObject<HTMLDivElement>;
 	look?: 'standard' | 'filled' | 'outlined' | string;
 	type: 'text' | 'password' | 'tel' | 'email' | 'hidden' | 'date';
@@ -36,7 +30,7 @@ interface InputTextProps extends Omit<InputHTMLAttributes<HTMLInputElement>, 'on
 	borderColor?: string;
 	useFloatingPlaceholder?: boolean;
 
-	icon?: NewIconProps[];
+	icon?: ICommon.NewIconProps[];
 
 	errorProps?: {};